Add git diff command to capture staged changes in CLI script
This commit adds a git diff --staged command to capture staged changes in a variable. The command is executed when the pre-commit hook is run. The staged changes are stored in the gitDiff variable, which can be used later in the script. A note has also been added to clarify that commits will already be staged before running the pre-commit hook.