Refactor git diff command to exclude lock files from staged files.
This commit updates the code in bin/cli.ts to exclude lock files from the git diff command that is run when a pre-commit hook is triggered. The commit uses an array of lock file names to generate exclude options, which are then passed to the git diff command. This is done to ensure that changes made to lock files are not included in the git commit.